Atualizado os identificadores das entidades e realizada uma possível solução para o problema do looping infinito do JSON

+++ patrimonio-back-end/src/main/java/br/edu/fatecsorocaba/system/endpoint/AcquisitionMethodEndpoint.java
@@ -37,14 +37,14 @@ public class AcquisitionMethodEndpoint {
 	}
 	
 	@GetMapping("/{id}")
-	public ResponseEntity<?> getById(@PathVariable("id") int id){
+	public ResponseEntity<?> getById(@PathVariable("id") Long id){
 		verifyIfacquisitionMethodExists(id);
 		Optional<AcquisitionMethod> acquisitionMethod = repository.findById(id);
 		return new ResponseEntity<>(acquisitionMethod, HttpStatus.OK);
 	}
 	
 	@DeleteMapping("/{id}")
-	public ResponseEntity<?> delete(@PathVariable("id") int id){
+	public ResponseEntity<?> delete(@PathVariable("id") Long id){
 		verifyIfacquisitionMethodExists(id);
 		repository.deleteById(id);
 		return new ResponseEntity<>(HttpStatus.OK);
@@ -57,7 +57,7 @@ public class AcquisitionMethodEndpoint {
 		return new ResponseEntity<>(HttpStatus.OK);
 	}
 	
-	public void verifyIfacquisitionMethodExists(int id) {
+	public void verifyIfacquisitionMethodExists(Long id) {
 		if(!repository.findById(id).isPresent())
 			throw new ResourceNotFoundException("acquisitionMethod with ID " + id + " not found.");
 	}
