+++ patrimonio-back-end/src/main/java/br/edu/fatecsorocaba/system/endpoint/LogEndpoint.java
@@ -37,14 +37,14 @@ public class LogEndpoint {
 	}
 	
 	@GetMapping("/{id}")
-	public ResponseEntity<?> getById(@PathVariable("id") int id){
+	public ResponseEntity<?> getById(@PathVariable("id") Long id){
 		verifyIflogExists(id);
 		Optional<Log> log = repository.findById(id);
 		return new ResponseEntity<>(log, HttpStatus.OK);
 	}
 	
 	@DeleteMapping("/{id}")
-	public ResponseEntity<?> delete(@PathVariable("id") int id){
+	public ResponseEntity<?> delete(@PathVariable("id") Long id){
 		verifyIflogExists(id);
 		repository.deleteById(id);
 		return new ResponseEntity<>(HttpStatus.OK);
@@ -57,7 +57,7 @@ public class LogEndpoint {
 		return new ResponseEntity<>(HttpStatus.OK);
 	}
 	
-	public void verifyIflogExists(int id) {
+	public void verifyIflogExists(Long id) {
 		if(!repository.findById(id).isPresent())
 			throw new ResourceNotFoundException("log with ID " + id + " not found.");
 	}
