+++ patrimonio-back-end/src/main/java/br/edu/fatecsorocaba/system/repository/AcquisitionMethodRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.edu.fatecsorocaba.system.model.AcquisitionMethod;
 
-public interface AcquisitionMethodRepository extends JpaRepository<AcquisitionMethod, Integer> {
+public interface AcquisitionMethodRepository extends JpaRepository<AcquisitionMethod, Long> {
 
 }
