+++ patrimonio-back-end/src/main/java/br/edu/fatecsorocaba/system/repository/LogRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.edu.fatecsorocaba.system.model.Log;
 
-public interface LogRepository extends JpaRepository<Log, Integer> {
+public interface LogRepository extends JpaRepository<Log, Long> {
 
 }
