+++ patrimonio-back-end/src/main/java/br/edu/fatecsorocaba/system/repository/PatrimonyRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.edu.fatecsorocaba.system.model.Patrimony;
 
-public interface PatrimonyRepository extends JpaRepository<Patrimony, Integer> {
+public interface PatrimonyRepository extends JpaRepository<Patrimony, Long> {
 
 }
